Rewrite cross to use dtolnay runner
As we do in other places stop using the `actions` runner and use the
`dtolnay` one to checkout toolchain.

While we are at it, use double quotes for `name` fields (this is a small
stylistic thing I have been introducing in an effort to make the yaml
files a bit easier to read).