musig: remove SessionSecretRand::new constructor
This is just a convenience wrapper around
SessionSecretRand::from_rng(&mut rng()). The name `new` disguises the
fact that this is an impure function and actually calls a RNG every time
it is called. It is also unused, in the docs or in the tests.

Finally, clippy complains about it because it feels that a `new`
method should always be accompanied by a `Default` impl...which is fair,
but the impurity is veen more surprising for `default` than for `new`.

Since `rand` renamed `thread_rng` to just `rng` the keystroke-saving
value of this method is less. So just delete it.