secp256k1-sys: fix `lowmemory` feature
When upstream switched to multicomb they changed the #define flags
needed to reduce the size of the ecmult_gen precomp table. We should
have updated our bulid.rs.

Before this change, on my system the secp256k1-sys rlib with the
lowmemory feature has size 630602. After this change, it has size
610090, a 3.3% reduction. Probably not worth backporting, although
it's not a breaking change and we totally could backport it.

Fixes #795