musig: Add warning to `SecretNonce::dangerous_into_bytes` about encoding instability
libsecp256k1 states that the byte format
`SecretNonce::dangerous_into_bytes` returns is not guaranteed to be
stable, and therefore rust-secp256k1 users must be careful to only pass
the result of `dangerous_into_bytes` to a compatible
`dangerous_from_bytes`, according to version and platform. The simplest
and most conservative strategy is to only consider same-versioned and
same-platform bytes as compatible. In practice, the format is unlikely
to change much if at all between most versions, so more permissive
strategies can be employed with research and great care by the user.