deprecations: replace all TBDs with 0.31.0
We accidentally released 0.30.0 and 0.31.0 with "future" deprecations.

Most users will not see these since the lint for them is different from
the normal deprecation warning (and it's undocumented except to say that
we "should not use" future deprecations, even though it works and there
is no alternative).

Somewhat fortunately, in 0.30.x we messed up the deprecations and
provided an inconsistent and weird API to replace the old functions. But
because we messed up the deprecation messages, we can pretend we didn't
do it and just mark the deprecations as having started in 0.31.x