context: introduce spinlock that gives up after a few iterations
Introduces a spinlocking mutex that only offers access to its internals
via a "try_unlock" method which spins a small finite number of times
before unlocking. We use a spinlock because, in the minimal dependency
set we support, there are no synchronization primitives except atomics,
so that's the only form of mutex we can create.

However, there are a number of problems with spinlocks -- see this article
(from Kix in #346) for some of them:

https://matklad.github.io/2020/01/02/spinlocks-considered-harmful.html

To avoid these problems, we give up after a few spins. The way we will
use this in the context object is:

1. When initializing the global context, if we can't get the lock, we
   just initialize a new stack-local context and use that. (A parallel
   thread must be initializing the context, which is wasteful but
   harmless.)
2. Once we unlock the context, we copy it onto the stack and re-lock it
   in order to minimize the time holding the lock. (The exception is
   during initialization where we hold the lock for the whole
   initialization, in the hopes that other threads will block on us
   instead of doing their own initialization.) If we rerandomize, we do
   this on the stack-local copy and then only re-lock to copy it back.
3. If we fail to get the lock to copy the rerandomized context back, we
   just don't copy it. The result is that we wasted some time
   rerandomizing without any benefit, which is not the end of the world.

The spinlock was implemented with help from ChatGPT o3 and the unit
tests with help from Claude 4 (though in both cases I did significant
refactoring and review by hand).