key.rs: clean up some pointer mangling in sort_pubkeys()
In sort_pubkeys() we convert a slice of PublicKeys to a slice of
ffi::PublicKeys. This is OK, because we have repr(transparent) allowing
us to do this. But we do so in an unnecessarily-complicated way, which
winds up going through the CPtr implementation on &[u8], and which I
think may technically be unsound.

Simplify it and avoid using this impl. The next commit will remove it
entirely.