minifest: Remove homepage
`homepage` is meaningless if it just points to the repo.

And from the docs:

A value should only be set for homepage if there is a dedicated
website for the crate other than the source repository or API
documentation. Do not make homepage redundant with either the
documentation or repository values.